requirements.txt
pytest~=7.4.3
allure-pytest~=2.13.2
requests~=2.31.0
pyleniumio~=1.20.0
Faker~=19.12.0